+++ BusBar/sql/dbmainele.h
@@ -6,7 +6,7 @@
 #define DB_TABLE_MAINELE "main_ele"
 
 struct DbMainEleItem : public DbBasicItem{
-    int line_1, line_2, line_3, line;
+    double line_1, line_2, line_3, line;
 };
